requirements.txt
azure-storage-blob
azure-eventhub
azure-eventhub-checkpointstoreblob-aio